requirements.txt
alembic==1.7.7
aniso8601==9.0.1
argon2-cffi==20.1.0
astroid==2.5.2
attrs==20.3.0
bcrypt==3.2.0
bidict==0.21.2
boto3==1.23.9
botocore==1.26.10
cachetools==4.2.1
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
click==7.1.2
exponent-server-sdk==0.3.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.1.0
Flask-RESTful==0.3.8
Flask-SocketIO==5.0.1
Flask-WTF==0.14.3
google-auth==1.28.0
gunicorn==20.1.0
idna==2.10
iniconfig==1.1.1
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
lazy-object-proxy==1.6.0
Mako==1.2.0
MarkupSafe==1.1.1
marshmallow==3.11.1
mccabe==0.6.1
numpy==1.22.4
packaging==20.9
pandas==1.4.2
pluggy==0.13.1
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.1
pylint==2.7.4
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
python-engineio==4.0.1
python-socketio==5.1.0
pytz==2021.1
requests==2.25.1
rsa==4.7.2
s3transfer==0.5.2
six==1.15.0
SQLAlchemy==1.3.5
toml==0.10.2
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
